# docs/swagger.yaml
definitions:
...
  Models.Location:
    properties:
      latitude:
        type: number
      longitude:
        type: number
      time:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: The worst IOT website in Laohe Mountain Institute of Technology
  title: IOTino
  version: 0.0.1
paths:
...
  /api/location/{device}:
    get:
      consumes:
      - application/json
      parameters:
      - description: device id
        in: path
        name: device
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/Models.Location'
            type: array
        "400":
          description: error
          schema:
            type: string
      summary: get certain device's history location sequence
      tags:
      - Location
  /api/locations:
    get:
      consumes:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/Models.Location'
            type: array
        "400":
          description: error
          schema:
            type: string
      summary: get all device's history location sequence
      tags:
      - Location
...
swagger: "2.0"